# docs/swagger.yaml
definitions:
  db.PriceLog:
    properties:
      coin:
        type: string
      id:
        type: integer
      priceUSD:
        type: number
      timestamp:
        type: string
    type: object
info:
  contact: {}
paths:
  /health:
    get:
      description: Returns OK if the service is running
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Service health check
      tags:
      - health
  /history:
    get:
      description: Returns N latest prices from DB
      parameters:
      - description: Coin ID
        in: query
        name: coin
        required: true
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.PriceLog'
            type: array
        "400":
          description: invalid request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Get coin price history
      tags:
      - query
  /latest:
    get:
      description: Returns the latest tracked price from DB
      parameters:
      - description: Coin ID
        in: query
        name: coin
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.PriceLog'
        "400":
          description: invalid request
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      summary: Get latest tracked coin price
      tags:
      - query
  /track:
    post:
      description: Fetches the current price from CoinGecko and saves to DB
      parameters:
      - description: Coin ID
        in: query
        name: coin
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.PriceLog'
        "400":
          description: invalid request
          schema:
            type: string
        "502":
          description: external API error
          schema:
            type: string
      summary: Track and save coin price
      tags:
      - tracking
swagger: "2.0"